Write tests for body helpers
