+++ src/core/http-body/__tests__/buffered-body.spec.ts
@@ -1,28 +1,21 @@
 import { BufferedBody } from "../buffered-body";
 import { stringToIterable } from "../string-encoding-utils";
-import { asJson, stringBody } from "../helpers";
 
-describe("HttpBodyImpl", () => {
-  it("asJson should return json", async () => {
-    const person = {
-      name: "John"
-    };
-    const body = stringBody(JSON.stringify(person));
-
-    expect(await asJson(body)).toEqual(person);
-  });
-
-  it("should return string when created from a string", async () => {
-    const content = "Some content";
-    const body = stringBody(content);
-
-    expect(await body.asString()).toEqual(content);
-  });
-
+describe("BufferedBody", () => {
   it("should correctly decode iterable contents when converting to string", async () => {
-    const it = stringToIterable("Hello");
+    const it = stringToIterable("Hello 😌");
     const body = new BufferedBody(it);
 
-    expect(await body.asString()).toEqual("Hello");
+    expect(await body.asString()).toEqual("Hello 😌");
+  });
+
+  it("should correctly decode string contents when converting to string", async () => {
+    async function* gen() {
+      yield "Hello 😌";
+    }
+
+    const body = new BufferedBody(gen());
+
+    expect(await body.asString()).toEqual("Hello 😌");
   });
 });
